Updated based on code review: * made new page for old migration guides * small fixes * moved inherit_doc in python
