+++ docs/ml-guide.md
@@ -174,8 +174,8 @@ val sqlContext = new SQLContext(sc)
 import sqlContext.implicits._
 
 // Prepare training data.
-// We use LabeledPoint, which is a case class.  Spark SQL can convert RDDs of Java Beans
-// into DataFrames, where it uses the bean metadata to infer the schema.
+// We use LabeledPoint, which is a case class.  Spark SQL can convert RDDs of case classes
+// into DataFrames, where it uses the case class metadata to infer the schema.
 val training = sc.parallelize(Seq(
   LabeledPoint(1.0, Vectors.dense(0.0, 1.1, 0.1)),
   LabeledPoint(0.0, Vectors.dense(2.0, 1.0, -1.0)),
