+++ python/pyspark/ml/feature.py
@@ -18,7 +18,7 @@
 from pyspark.ml.param.shared import HasInputCol, HasOutputCol, HasNumFeatures
 from pyspark.ml.util import keyword_only
 from pyspark.ml.wrapper import JavaTransformer
-from pyspark.mllib.__init__ import inherit_doc
+from pyspark.mllib.common import inherit_doc
 
 __all__ = ['Tokenizer', 'HashingTF']
