+++ python/pyspark/mllib/__init__.py
@@ -21,7 +21,6 @@ Python bindings for MLlib.
 
 # MLlib currently needs NumPy 1.4+, so complain if lower
 
-import inspect
 import numpy
 if numpy.version.version < '1.4':
     raise Exception("MLlib requires NumPy 1.4+")
@@ -35,19 +34,3 @@ random.__name__ = 'random'
 random.RandomRDDs.__module__ = __name__ + '.random'
 sys.modules[__name__ + '.random'] = random
 
-
-def inherit_doc(cls):
-    """
-    A decorator that makes a class inherit documentation from its parents.
-    """
-    for name, func in vars(cls).items():
-        # only inherit docstring for public functions
-        if name.startswith("_"):
-            continue
-        if not func.__doc__:
-            for parent in inspect.getmro(cls):
-                parent_func = getattr(parent, name, None)
-                if parent_func and getattr(parent_func, "__doc__", None):
-                    func.__doc__ = parent_func.__doc__
-                    break
-    return cls
